refactor: consolidate Sleeper sport handlers into shared factory
Football and basketball handler files were identical except for the
sport state path (/state/nfl vs /state/nba) and player cache sport key.
Extract shared logic into createSleeperHandlers() factory, reducing
~770 lines to ~380 + 2 thin sport wrappers (~3 lines each).

Adds 10 cross-sport characterization tests covering get_league_info,
get_standings, get_roster, and get_matchups across both sports, including
verification that the correct sport-specific state endpoint is called.
